requirements.txt
streamlit
diffusers
Pillow
googletrans